requirements.txt
numpy
websocket-client
shapely
